Hiding object detection for now.
Former-commit-id: 095504fa43ad936455ab5c9f49e6c5311d869dc8
Former-commit-id: 2c2b3e977e3a37291efad501d8e8585d8ce5fac9
Former-commit-id: f73c25ca0155175e53d415c325757443dc237f95

+++ src/ImageProcessor/Imaging/Quantizers/OctreeQuantizer.cs
@@ -506,12 +506,10 @@ namespace ImageProcessor.Imaging.Quantizers
                     {
                         // Consume the next palette index
                         this.paletteIndex = index++;
-                        //int r = Math.Abs(this.red / this.pixelCount);
-                        //int g = Math.Abs(this.green / this.pixelCount);
-                        //int b = Math.Abs(this.blue / this.pixelCount);
-                        int r = (this.red / this.pixelCount).ToByte();
-                        int g = (this.green / this.pixelCount).ToByte();
-                        int b = (this.blue / this.pixelCount).ToByte();
+
+                        byte r = (this.red / this.pixelCount).ToByte();
+                        byte g = (this.green / this.pixelCount).ToByte();
+                        byte b = (this.blue / this.pixelCount).ToByte();
 
                         // And set the color of the palette entry
                         palette.Add(Color.FromArgb(r, g, b));
